+++ torchtitan/experiments/graph_trainer/passes.py
@@ -25,20 +25,12 @@ from collections.abc import Callable
 
 import torch
 from torch._inductor.compile_fx import compile_fx_inner
-from torch._inductor.fx_passes.bucketing import (
-    is_all_gather_into_tensor as is_all_gather,
-)
-from torch._inductor.fx_passes.overlap_manual_scheduling import manual_overlap_bucketing
-from torch._inductor.fx_passes.overlap_scheduling import schedule_overlap_bucketing
 from torch._logging import trace_structured
 from torch.fx.passes.regional_inductor import regional_inductor
 from torch.utils.checkpoint import CheckpointPolicy
 
 from torchtitan.distributed.activation_checkpoint import _get_save_ops
 from torchtitan.distributed.fsdp import get_fsdp_reshard_after_forward_policy
-from torchtitan.experiments.graph_trainer.bucketing import (
-    joint_transformer_block_bucketing_reordering_pass,
-)
 from torchtitan.experiments.graph_trainer.common_utils import (
     _get_layer_id,
     _is_backward_node,
@@ -53,6 +45,13 @@ from torchtitan.experiments.graph_trainer.debug_utils import (
     log_graph_diff,
     snapshot_graph,
 )
+from torchtitan.experiments.graph_trainer.fsdp_passes import (
+    autobucketing_reordering_pass,
+    fsdp_reshard_after_fwd_pass,
+    joint_transformer_block_bucketing_reordering_pass,
+    overlap_fsdp_ag_rs_pass,
+    transformer_block_bucketing_reordering_pass,
+)
 from torchtitan.experiments.graph_trainer.log_activation_memory_policy import (
     log_activation_memory_policy,
 )
@@ -61,10 +60,6 @@ from torchtitan.experiments.graph_trainer.remove_noop_passes import (
     remove_detach_pass,
     remove_identity_slice_pass,
     remove_identity_view_pass,
-)
-from torchtitan.experiments.graph_trainer.reshard_after_forward import (
-    annotate_fsdp_all_gather,
-    is_wait_tensor_from_fsdp,
 )
 from torchtitan.tools.logging import logger
 
@@ -302,36 +297,6 @@ def apply_graph_passes(
             tlparse_log_graph_pass(gm, graph_name=f"after_{pass_name}", debug=debug)
             after_snapshot = snapshot_graph(gm)
             log_graph_diff(before_snapshot, after_snapshot, pass_name)
-    return gm
-
-
-def autobucketing_reordering_pass(
-    gm: torch.fx.GraphModule, example_inputs: tuple | None = None
-) -> torch.fx.GraphModule:
-    """
-    Apply autobucketing and reordering optimization.
-
-    This pass applies schedule_overlap_bucketing with collective_bucketing enabled
-    to optimize comm/compute overlap patterns in the graph.
-    """
-    schedule_overlap_bucketing(gm, collective_bucketing=True)
-    gm.recompile()
-    return gm
-
-
-def transformer_block_bucketing_reordering_pass(
-    gm: torch.fx.GraphModule,
-    example_inputs: tuple | None = None,
-    *,
-    fsdp_manual_buckets,
-) -> torch.fx.GraphModule:
-    """
-    Apply aten-level manual bucketing and reordering optimization.
-    """
-    manual_overlap_bucketing(
-        gm, module_bucket_plans=fsdp_manual_buckets, insert_overlap_deps=False
-    )
-    gm.recompile()
     return gm
 
 
@@ -866,23 +831,6 @@ def selective_activation_remat_pass(
     return remat_using_tags_for_fwd_loss_bwd_graph(gm)
 
 
-# Apply activation checkpointing on joint graph before partitioner
-def fsdp_reshard_after_fwd_pass(
-    gm: torch.fx.GraphModule,
-    example_inputs: tuple | None = None,
-    *,
-    reshard_after_forward: bool,
-) -> torch.fx.GraphModule:
-    # this pass implements simplefsdp's fsdp_reshard_after_forward behavior
-    # when fsdp_reshard_after_forward set to True, it will annotate simple_fsdp AG
-    #   to CheckpointPolicy.MUST_RECOMPUTE.
-    # when fsdp_reshard_after_forward set to False, it will annotate simple_fsdp AG
-    #   to CheckpointPolicy.MUST_SAVE.
-    gm = annotate_fsdp_all_gather(gm, reshard_after_forward)
-    gm.recompile()
-    return gm
-
-
 def full_inductor_compilation_pass(
     gm: torch.fx.GraphModule, example_inputs: tuple
 ) -> torch.fx.GraphModule:
@@ -935,80 +883,6 @@ def full_inductor_compilation_pass(
         return output_code(list(args))
 
     gm.forward = _compiled_forward
-    return gm
-
-
-# Maps an FSDP group_name to an extra group_name created by this pass.
-# Each NCCL PG gets its own CUDA stream, so the extra PG is what enables
-# AG/RS overlap in backward.
-_EXTRA_FSDP_PG_REGISTRY: dict[str, str] = {}
-
-
-def _get_or_create_extra_fsdp_pg(source_pg_name: str) -> str:
-    """Return the extra PG name for ``source_pg_name``, creating it once.
-
-    The extra PG is a new NCCL process group with the same ranks as the source
-    FSDP PG but a different communicator (and therefore a different CUDA stream).
-    """
-    import torch.distributed as dist
-
-    if source_pg_name in _EXTRA_FSDP_PG_REGISTRY:
-        return _EXTRA_FSDP_PG_REGISTRY[source_pg_name]
-
-    source_pg = dist.distributed_c10d._resolve_process_group(source_pg_name)
-    ranks = dist.get_process_group_ranks(source_pg)
-    extra_pg = dist.new_group(
-        ranks=ranks, group_desc="fsdp_extra", use_local_synchronization=True
-    )
-    _EXTRA_FSDP_PG_REGISTRY[source_pg_name] = extra_pg.group_name
-    logger.info(
-        f"Created extra FSDP PG (source: {source_pg_name}, "
-        f"extra: {extra_pg.group_name})"
-    )
-    return extra_pg.group_name
-
-
-def overlap_fsdp_ag_rs_pass(
-    gm: torch.fx.GraphModule,
-    example_inputs: tuple | None = None,
-) -> torch.fx.GraphModule:
-    """
-    Reassign FSDP all-gather nodes to an extra NCCL process group for
-    AG/RS overlap in backward.
-
-    Discovers the FSDP PG by inspecting the graph, creates an extra
-    NCCL PG over the same ranks (giving it a separate CUDA stream),
-    and rewrites every all-gather using that source PG to the extra PG.
-    This separates all-gathers from reduce-scatters onto different streams,
-    enabling AG/RS overlap in backward.
-
-    No-op when the graph has no FSDP all-gathers. Must be applied BEFORE
-    bucketing passes so bucketed all-gathers inherit the new PG name.
-    """
-    source_pg_name: str | None = None
-    for node in gm.graph.nodes:
-        if is_wait_tensor_from_fsdp(node):
-            ag_node = node.args[0]
-            source_pg_name = ag_node.args[2]
-            break
-
-    if source_pg_name is None:
-        return gm
-
-    target_pg_name = _get_or_create_extra_fsdp_pg(source_pg_name)
-
-    count = 0
-    for node in gm.graph.nodes:
-        if is_all_gather(node) and node.args[2] == source_pg_name:
-            # AG args: (input_tensor, group_size, group_name)
-            node.args = (node.args[0], node.args[1], target_pg_name)
-            count += 1
-    if count > 0:
-        logger.info(
-            f"Rewrote {count} all-gather node(s) from PG {source_pg_name} "
-            f"to PG {target_pg_name}"
-        )
-    gm.recompile()
     return gm
