+++ torchtitan/experiments/graph_trainer/tests/test_passes.py
@@ -26,6 +26,7 @@ from torchtitan.experiments.graph_trainer.common_utils import (
     _MODULE_FQN,
     annotate_module_fqns,
 )
+from torchtitan.experiments.graph_trainer.fsdp_passes import overlap_fsdp_ag_rs_pass
 from torchtitan.experiments.graph_trainer.graph_utils import export_joint
 from torchtitan.experiments.graph_trainer.make_fx_tracer import (
     minimal_fx_tracer,
@@ -35,7 +36,6 @@ from torchtitan.experiments.graph_trainer.passes import (
     _make_default_memory_policy,
     apply_sac_pass,
     insert_kernel_annotations_pass,
-    overlap_fsdp_ag_rs_pass,
     remove_detach_pass,
     remove_identity_slice_pass,
     remove_identity_view_pass,
@@ -145,7 +145,9 @@ class TestOverlapFsdpAgRsPass(FSDPTest):
     def test_overlap_rewrites_ag_nodes(self):
         """Apply overlap_fsdp_ag_rs_pass on the real backward graph and verify
         that FSDP AG nodes are rewritten to the auto-created extra PG."""
-        from torchtitan.experiments.graph_trainer.passes import _EXTRA_FSDP_PG_REGISTRY
+        from torchtitan.experiments.graph_trainer.fsdp_passes import (
+            _EXTRA_FSDP_PG_REGISTRY,
+        )
 
         self._setup()
         model = self._make_fsdp_model()
